requirements.txt
requests
beautifulsoup4
aiohttp
rich
typer
PyQt6
cloudscraper
Pillow
